Optimize unit tests
