+++ view.go
@@ -44,6 +44,8 @@ func (receive *View) First(views []string, data ...any) {
 			return
 		}
 	}
+
+	panic("no view exists")
 }
 
 func structToMap(data any) map[string]any {
@@ -57,6 +59,9 @@ func structToMap(data any) map[string]any {
 	}
 
 	for i := 0; i < modelType.NumField(); i++ {
+		if !modelType.Field(i).IsExported() {
+			continue
+		}
 		dbColumn := modelType.Field(i).Name
 		if modelValue.Field(i).Kind() == reflect.Pointer {
 			if modelValue.Field(i).IsNil() {
